Protecting Task from wrong user manipulation

+++ backend/controller/tasksController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require('express-async-handler')
 const Task = require('../model/taskModel');
+const User = require('../model/userModel');
 
 const getTasks = asyncHandler(async(req,res) => {
 
@@ -23,12 +24,45 @@ const updateTask = asyncHandler (async(req, res) => {
         res.status(400);
         throw new Error('Task not found');
     }
+
+    const user = await User.findById(req.user.id);
+
+    if(!user){
+        res.status(401);
+        throw new Error('No such user found');
+    }
+
+    if(task.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User is not authorized to update');
+    }
+
     const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true})
     res.status(200).json(updatedTask);
 })
 
 const deleteTask = asyncHandler (async (req, res) => {
-    const task = await Task.findByIdAndDelete(req.params.id);
+    const task = await Task.findById(req.params.id);
+
+    if(!task){
+        res.status(400);
+        throw new Error('Task not found')
+    }
+
+
+    const user = await User.findById(req.user.id);
+
+    if(!user){
+        res.status(401);
+        throw new Error('No such user found');
+    }
+
+    if(task.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User is not authorized to update');
+    }
+
+    task = await Task.findByIdAndDelete(req.params.id);
     res.status(200).json({message: `Deleted Task ${req.params.id}`});
 })
